P2a: extract afxdp/tx/stats.rs + plan v4
- afxdp.rs:99 #[path] updated tx.rs → tx/mod.rs (prerequisite for move).
- tx.rs renamed to tx/mod.rs.
- New tx/stats.rs with stamp_submits, record_kick_latency,
  record_tx_completions_with_stamp + their doc blocks (~155 LOC).
- tx/mod.rs adds 'pub(super) mod stats;' + load-bearing
  pub(in crate::afxdp) re-export so umem::tests at umem.rs:950+ keep
  resolving crate::afxdp::tx::{stamp_submits,
  record_tx_completions_with_stamp, record_kick_latency}.
- Two record_* fns visibility bumped from pub(super) to
  pub(in crate::afxdp).
- UMEM_FRAME_SHIFT imported from crate::afxdp:: (afxdp.rs:162) since
  tx/mod.rs no longer inherits it via 'use super::*;' for the moved
  fns' new home.

cargo build --bins clean (88 warnings, all pre-existing).
cargo test --bins 865/0/2 (rolling baseline).

Plan: docs/pr/p2a-tx-stats/plan.md (v4, addressing Codex r1+r2+cross
findings).
